+++ examples/client/processor/test.hpp
@@ -1,0 +1,13 @@
+#pragma once
+#include <memory>
+
+#include <streamline/inspector.hpp>
+#include <streamline/system.hpp>
+
+#include "config.hpp"
+
+namespace scheduler {
+class Scheduler;
+}
+
+void test_outputer(scheduler::Scheduler& scheduler, OutputConfig const& output);