+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [4.0.2] 2025-03-12
+
+- TRACE/VERBOSE/DEBUG is now available in Release mode
+- Added `--output-print-everything` option to set/override `print=true` for every output
+- Added experimental `--output tcp-server:` for outputing as a TCP server to all connected clients.
+- Added new output format `test`. The test format will send the string `TESTTESTTESTTEST` every second to all outputs that includes the `test` format. This is useful for testing if a output is working as expected. 
+
 ## [4.0.1] 2025-03-07
 
 Due to extensive architectural changes and improvements between versions 3.x and 4.0, a comprehensive changelog was not maintained. The upgrade includes significant modifications to core functionality and structure. Starting from version 4.0, all changes will be thoroughly documented in this changelog.